+++ tests/integration/executor/test_environment_overlay.py
@@ -28,8 +28,27 @@ def setup_function():
     callbacks.unregister_all()
 
 
+def setup_module():
+    craft_parts.Features.reset()
+    craft_parts.Features(enable_overlay=True)
+
+
 def teardown_module():
     callbacks.unregister_all()
+    craft_parts.Features.reset()
+
+
+@pytest.fixture(autouse=True)
+def mock_mount_unmount(mocker):
+    mocker.patch("craft_parts.utils.os_utils.mount")
+    mocker.patch("craft_parts.utils.os_utils.mount_overlayfs")
+    mocker.patch("craft_parts.utils.os_utils.umount")
+
+
+@pytest.fixture(autouse=True)
+def mock_prerequisites_for_overlay(mocker):
+    mocker.patch("craft_parts.lifecycle_manager._ensure_overlay_supported")
+    mocker.patch("craft_parts.overlays.OverlayManager.refresh_packages_list")
 
 
 def _prologue_callback(info: ProjectInfo) -> None:
@@ -52,6 +71,7 @@ _parts_yaml = textwrap.dedent(
       foo:
         plugin: nil
         override-pull: env | egrep "^(TEST|CRAFT)_" | sort
+        overlay-script: env | egrep "^(TEST|CRAFT)_" | sort
         override-build: env | egrep "^(TEST|CRAFT)_" | sort
         override-stage: env | egrep "^(TEST|CRAFT)_" | sort
         override-prime: env | egrep "^(TEST|CRAFT)_" | sort
@@ -61,7 +81,7 @@ _parts_yaml = textwrap.dedent(
 )
 
 
-@pytest.mark.parametrize("step", [Step.PULL, Step.BUILD, Step.STAGE, Step.PRIME])
+@pytest.mark.parametrize("step", list(Step))
 def test_step_callback(new_dir, mocker, capfd, step):
     mocker.patch("platform.machine", return_value="aarch64")
 
@@ -85,6 +105,7 @@ def test_step_callback(new_dir, mocker, capfd, step):
         textwrap.dedent(
             f"""\
             CRAFT_ARCH_TRIPLET=aarch64-linux-gnu
+            CRAFT_OVERLAY={new_dir}/overlay/overlay
             CRAFT_PARALLEL_BUILD_COUNT=1
             CRAFT_PART_BUILD={new_dir}/parts/foo/build
             CRAFT_PART_BUILD_WORK={new_dir}/parts/foo/build
@@ -127,6 +148,7 @@ def test_prologue_callback(new_dir, capfd, mocker):
         textwrap.dedent(
             f"""\
             CRAFT_ARCH_TRIPLET=aarch64-linux-gnu
+            CRAFT_OVERLAY={new_dir}/overlay/overlay
             CRAFT_PARALLEL_BUILD_COUNT=1
             CRAFT_PART_BUILD={new_dir}/parts/foo/build
             CRAFT_PART_BUILD_WORK={new_dir}/parts/foo/build
