+++ tests/integration/lifecycle/test_lifecycle_overlay.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright 2021-2023 Canonical Ltd.
+# Copyright 2021 Canonical Ltd.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -21,7 +21,7 @@ import pytest
 import yaml
 
 import craft_parts
-from craft_parts import Action, ActionProperties, ActionType, Features, Step
+from craft_parts import Action, ActionProperties, ActionType, Step
 
 basic_parts_yaml = textwrap.dedent(
     """\
@@ -41,10 +41,10 @@ basic_parts_yaml = textwrap.dedent(
 
 @pytest.fixture(autouse=True)
 def _setup_module():
-    Features.reset()
-    Features(enable_overlay=False)
+    craft_parts.Features.reset()
+    craft_parts.Features(enable_overlay=True)
     yield
-    Features.reset()
+    craft_parts.Features.reset()
 
 
 def test_basic_lifecycle_actions(new_dir, mocker):
@@ -80,6 +80,11 @@ def test_basic_lifecycle_actions(new_dir, mocker):
     assert actions == [
         # fmt: off
         Action("foobar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", Step.OVERLAY, action_type=ActionType.RUN, reason="required to overlay 'foobar'"),
+        Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", Step.OVERLAY, action_type=ActionType.RUN, reason="required to overlay 'foobar'"),
+        Action("foobar", Step.OVERLAY),
         Action("foobar", Step.BUILD),
         Action("foobar", Step.STAGE),
         Action("foobar", Step.PRIME),
@@ -96,7 +101,9 @@ def test_basic_lifecycle_actions(new_dir, mocker):
     actions = lf.plan(Step.BUILD, ["bar"])
     assert actions == [
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.BUILD, reason="required to build 'bar'"),
         Action("foo", Step.STAGE, reason="required to build 'bar'"),
         Action("bar", Step.BUILD),
@@ -112,6 +119,7 @@ def test_basic_lifecycle_actions(new_dir, mocker):
     assert actions == [
         # fmt: off
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("bar", Step.BUILD, action_type=ActionType.RERUN, reason="requested step"),
         # fmt: on
     ]
@@ -129,7 +137,9 @@ def test_basic_lifecycle_actions(new_dir, mocker):
     assert actions == [
         # fmt: off
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.PULL, action_type=ActionType.RERUN, reason="'source' property changed"),
+        Action("foo", Step.OVERLAY, action_type=ActionType.RUN, reason="required to build 'bar'"),
         Action("foo", Step.BUILD, action_type=ActionType.RUN, reason="required to build 'bar'"),
         Action("foo", Step.STAGE, action_type=ActionType.RUN, reason="required to build 'bar'"),
         Action("bar", Step.BUILD, action_type=ActionType.RERUN, reason="requested step"),
@@ -148,6 +158,9 @@ def test_basic_lifecycle_actions(new_dir, mocker):
         Action("foo", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
         Action("foobar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foobar", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
         Action("bar", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
         Action("foobar", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
@@ -166,9 +179,13 @@ def test_basic_lifecycle_actions(new_dir, mocker):
                properties=ActionProperties(changed_files=['a.tar.gz'], changed_dirs=[])),
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
         Action("foobar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", step=Step.OVERLAY, action_type=ActionType.UPDATE, reason="'PULL' step changed"),
+        Action("bar", step=Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foobar", step=Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.BUILD, action_type=ActionType.UPDATE, reason="'PULL' step changed",
                properties=ActionProperties(changed_files=['a.tar.gz'], changed_dirs=[])),
         Action("foo", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.STAGE, action_type=ActionType.RERUN, reason="required to build 'bar'"),
         Action("bar", Step.BUILD, action_type=ActionType.RERUN, reason="stage for part 'foo' changed"),
@@ -185,6 +202,9 @@ def test_basic_lifecycle_actions(new_dir, mocker):
         Action("foo", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
         Action("bar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
         Action("foobar", Step.PULL, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foo", step=Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
+        Action("bar", step=Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
+        Action("foobar", step=Step.OVERLAY, action_type=ActionType.SKIP, reason="already ran"),
         Action("foo", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
         Action("bar", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
         Action("foobar", Step.BUILD, action_type=ActionType.SKIP, reason="already ran"),
@@ -261,6 +281,8 @@ class TestCleaning:
 
         assert sorted(state_dir.rglob("*")) == [
             state_dir / "build",
+            state_dir / "layer_hash",
+            state_dir / "overlay",
             state_dir / "prime",
             state_dir / "pull",
             state_dir / "stage",
@@ -297,6 +319,8 @@ class TestCleaning:
         assert Path("prime/bar.txt").is_file()
         assert sorted(bar_state_dir.rglob("*")) == [
             bar_state_dir / "build",
+            bar_state_dir / "layer_hash",
+            bar_state_dir / "overlay",
             bar_state_dir / "prime",
             bar_state_dir / "pull",
             bar_state_dir / "stage",
@@ -314,6 +338,7 @@ class TestCleaning:
         bar_state_dir = Path("parts/bar/state")
 
         step_is_overlay_or_later = step >= Step.OVERLAY
+        step_is_build_or_later = step >= Step.BUILD
         step_is_stage_or_later = step >= Step.STAGE
         step_is_prime = step == Step.PRIME
 
@@ -332,6 +357,11 @@ class TestCleaning:
         if step_is_overlay_or_later:
             all_states.append(foo_state_dir / "pull")
             all_states.append(bar_state_dir / "pull")
+        if step_is_build_or_later:
+            all_states.append(foo_state_dir / "overlay")
+            all_states.append(bar_state_dir / "overlay")
+            all_states.append(foo_state_dir / "layer_hash")
+            all_states.append(bar_state_dir / "layer_hash")
         if step_is_stage_or_later:
             all_states.append(foo_state_dir / "build")
             all_states.append(bar_state_dir / "build")
