+++ tests/unit/executor/test_part_handler.py
@@ -27,7 +27,6 @@ from craft_parts.actions import Action, ActionType
 from craft_parts.executor import filesets, part_handler
 from craft_parts.executor.part_handler import PartHandler
 from craft_parts.executor.step_handler import StepContents
-from craft_parts.features import Features
 from craft_parts.infos import PartInfo, ProjectInfo, StepInfo
 from craft_parts.overlays import OverlayManager
 from craft_parts.parts import Part
@@ -36,15 +35,6 @@ from craft_parts.steps import Step
 from craft_parts.utils import os_utils
 
 # pylint: disable=too-many-lines
-
-
-def setup_module():
-    Features.reset()
-    Features(enable_overlay=True)
-
-
-def teardown_module():
-    Features.reset()
 
 
 @pytest.mark.usefixtures("new_dir")
@@ -62,12 +52,11 @@ class TestPartHandling:
                 "stage-packages": ["pkg1"],
                 "stage-snaps": ["snap1"],
                 "build-packages": ["pkg3"],
-                "overlay-packages": ["pkg4"],
             },
         )
         info = ProjectInfo(application_name="test", cache_dir=new_dir)
         ovmgr = OverlayManager(
-            project_info=info, part_list=[self._part], base_layer_dir=Path("/base")
+            project_info=info, part_list=[self._part], base_layer_dir=None
         )
         self._part_info = PartInfo(info, self._part)
         self._handler = PartHandler(
@@ -106,44 +95,6 @@ class TestPartHandling:
             },
         )
 
-    def test_run_overlay(self, mocker):
-        mocker.patch("craft_parts.overlays.OverlayManager.download_packages")
-        mocker.patch("craft_parts.overlays.OverlayManager.install_packages")
-
-        state = self._handler._run_overlay(
-            StepInfo(self._part_info, Step.OVERLAY), stdout=None, stderr=None
-        )
-        assert state == states.OverlayState(
-            part_properties=self._part.spec.marshal(),
-            project_options=self._part_info.project_options,
-        )
-
-    def test_run_overlay_with_filter(self, mocker, new_dir):
-        mocker.patch("craft_parts.overlays.OverlayManager.download_packages")
-        mocker.patch("craft_parts.overlays.OverlayManager.install_packages")
-
-        p1 = Part("p1", {"plugin": "nil", "overlay": ["-foo"]})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1], base_layer_dir=Path("/base")
-        )
-        part_info = PartInfo(info, p1)
-        handler = PartHandler(
-            p1, part_info=part_info, part_list=[p1], overlay_manager=ovmgr
-        )
-
-        p1.part_layer_dir.mkdir(parents=True)
-        file1 = p1.part_layer_dir / "foo"
-        file2 = p1.part_layer_dir / "bar"
-
-        file1.touch()
-        file2.touch()
-
-        handler._run_overlay(StepInfo(part_info, Step.PULL), stdout=None, stderr=None)
-
-        assert file1.exists() is False
-        assert file2.is_file()
-
     def test_run_build(self, mocker):
         mocker.patch("craft_parts.executor.step_handler.StepHandler._builtin_build")
         mocker.patch(
@@ -169,18 +120,11 @@ class TestPartHandling:
                 "installed-packages": ["hello=2.10"],
                 "installed-snaps": ["snapcraft=6466"],
             },
-            overlay_hash="d12e3f53ba91f94656abc940abb50b12b209d246",
+            overlay_hash="6554e32fa718d54160d0511b36f81458e4cb2357",
         )
 
-        self._mock_mount_overlayfs.assert_called_with(
-            f"{self._part_info.overlay_mount_dir}",
-            (
-                f"-olowerdir={self._part_info.overlay_packages_dir}:/base,"
-                f"upperdir={self._part.part_layer_dir},"
-                f"workdir={self._part_info.overlay_work_dir}"
-            ),
-        )
-        self._mock_umount.assert_called_with(f"{self._part_info.overlay_mount_dir}")
+        assert self._mock_mount_overlayfs.mock_calls == []
+        assert self._mock_umount.mock_calls == []
 
     @pytest.mark.usefixtures("new_dir")
     @pytest.mark.parametrize("out_of_source", [True, False])
@@ -220,9 +164,7 @@ class TestPartHandling:
 
         p1 = Part("p1", {"plugin": "nil"})
         info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1], base_layer_dir=Path("/base")
-        )
+        ovmgr = OverlayManager(project_info=info, part_list=[p1], base_layer_dir=None)
         part_info = PartInfo(info, p1)
         handler = PartHandler(
             p1, part_info=part_info, part_list=[p1], overlay_manager=ovmgr
@@ -245,7 +187,7 @@ class TestPartHandling:
             project_options=self._part_info.project_options,
             files={"file"},
             directories={"dir"},
-            overlay_hash="d12e3f53ba91f94656abc940abb50b12b209d246",
+            overlay_hash="6554e32fa718d54160d0511b36f81458e4cb2357",
         )
 
     def test_run_prime(self, new_dir, mocker):
@@ -257,7 +199,7 @@ class TestPartHandling:
 
         info = ProjectInfo(application_name="test", cache_dir=new_dir)
         ovmgr = OverlayManager(
-            project_info=info, part_list=[self._part], base_layer_dir=Path("/base")
+            project_info=info, part_list=[self._part], base_layer_dir=None
         )
         handler = PartHandler(
             self._part,
@@ -300,7 +242,6 @@ class TestPartHandling:
         "step,scriptlet",
         [
             (Step.PULL, "override-pull"),
-            (Step.OVERLAY, "overlay-script"),
             (Step.BUILD, "override-build"),
             (Step.STAGE, "override-stage"),
             (Step.PRIME, "override-prime"),
@@ -406,7 +347,6 @@ class TestPartHandling:
         "step,scriptlet",
         [
             (Step.PULL, "override-pull"),
-            (Step.OVERLAY, "overlay-script"),
             (Step.BUILD, "override-build"),
             (Step.STAGE, "override-stage"),
             (Step.PRIME, "override-prime"),
@@ -441,36 +381,6 @@ class TestPartHandling:
         assert err == ""
         assert output_path.read_text() == "hello\n"
         assert error_path.read_text() == "+ echo hello\n+ echo goodbye\ngoodbye\n"
-
-    def test_compute_layer_hash(self, new_dir):
-        p1 = Part("p1", {"plugin": "nil", "overlay-packages": ["pkg1"]})
-        p2 = Part("p2", {"plugin": "nil", "overlay-script": "ls"})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        part_info = PartInfo(info, p1)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1, p2], base_layer_dir=None
-        )
-        handler = PartHandler(
-            p1, part_info=part_info, part_list=[p1, p2], overlay_manager=ovmgr
-        )
-
-        layer_hash = handler._compute_layer_hash(all_parts=False)
-        assert layer_hash.hex() == "80ab51c6c76eb2b6fc01adc3143ebaf2b982ae56"
-
-    def test_compute_layer_hash_for_all_parts(self, new_dir):
-        p1 = Part("p1", {"plugin": "nil", "overlay-packages": ["pkg1"]})
-        p2 = Part("p2", {"plugin": "nil", "overlay-script": "ls"})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        part_info = PartInfo(info, p1)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1, p2], base_layer_dir=None
-        )
-        handler = PartHandler(
-            p1, part_info=part_info, part_list=[p1, p2], overlay_manager=ovmgr
-        )
-
-        layer_hash = handler._compute_layer_hash(all_parts=True)
-        assert layer_hash.hex() == "f4ae5a2ed1b4fd8a7e03f9264ab0f98ed6fd991b"
 
 
 @pytest.mark.usefixtures("new_dir")
@@ -606,271 +516,6 @@ class TestPartUpdateHandler:
     def test_update_invalid(self, step):
         with pytest.raises(errors.InvalidAction):
             self._handler.run_action(Action("foo", step, ActionType.UPDATE))
-
-
-@pytest.mark.usefixtures("new_dir")
-class TestPartReapplyHandler:
-    """Verify step reapplication processing."""
-
-    @pytest.fixture(autouse=True)
-    def setup_method_fixture(self, mocker, new_dir):
-        # pylint: disable=attribute-defined-outside-init
-        self._part = Part("foo", {"plugin": "nil", "overlay-script": "touch bar.txt"})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[self._part], base_layer_dir=new_dir
-        )
-        self._part_info = PartInfo(info, self._part)
-        self._handler = PartHandler(
-            self._part,
-            part_info=self._part_info,
-            part_list=[self._part],
-            overlay_manager=ovmgr,
-        )
-
-        mocker.patch("craft_parts.utils.os_utils.mount")
-        mocker.patch("craft_parts.utils.os_utils.mount_overlayfs")
-        mocker.patch("craft_parts.utils.os_utils.umount")
-        # pylint: enable=attribute-defined-outside-init
-
-    def test_reapply_overlay(self):
-        self._handler.run_action(Action("foo.txt", Step.PULL))
-
-        Path("parts/foo/layer/foo").touch()
-
-        self._handler.run_action(Action("foo", Step.OVERLAY, ActionType.REAPPLY))
-
-        assert Path("parts/foo/layer/foo.txt").exists() is False
-        assert Path("parts/foo/layer/bar.txt").exists()
-
-    @pytest.mark.parametrize("step", [Step.PULL, Step.BUILD, Step.STAGE, Step.PRIME])
-    def test_reapply_invalid(self, step):
-        with pytest.raises(errors.InvalidAction):
-            self._handler.run_action(Action("foo", step, ActionType.REAPPLY))
-
-
-@pytest.mark.usefixtures("new_dir")
-class TestOverlayMigration:
-    """Overlay migration to stage and prime test cases"""
-
-    @pytest.fixture(autouse=True)
-    def setup_method_fixture(self, new_dir):
-        # pylint: disable=attribute-defined-outside-init
-        p1 = Part("p1", {"plugin": "nil", "overlay-script": "ls"})
-        p2 = Part("p2", {"plugin": "nil", "overlay-script": "ls"})
-        p3 = Part("p3", {"plugin": "nil"})
-
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1, p2, p3], base_layer_dir=None
-        )
-
-        self._p1_handler = PartHandler(
-            p1,
-            part_info=PartInfo(info, p1),
-            part_list=[p1, p2, p3],
-            overlay_manager=ovmgr,
-        )
-        self._p2_handler = PartHandler(
-            p2,
-            part_info=PartInfo(info, p2),
-            part_list=[p1, p2, p3],
-            overlay_manager=ovmgr,
-        )
-        self._p3_handler = PartHandler(
-            p3,
-            part_info=PartInfo(info, p3),
-            part_list=[p1, p2, p3],
-            overlay_manager=ovmgr,
-        )
-
-        self._p1_handler._make_dirs()
-        self._p2_handler._make_dirs()
-        self._p3_handler._make_dirs()
-
-        # populate layers
-        Path(p1.part_layer_dir, "dir1").mkdir()
-        Path(p1.part_layer_dir, "dir1/foo").touch()
-        Path(p1.part_layer_dir, "bar").touch()
-
-        Path(p2.part_layer_dir, "dir1").mkdir()
-        Path(p2.part_layer_dir, "dir1/baz").touch()
-        # pylint: enable=attribute-defined-outside-init
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_migrate_overlay(self, step, step_dir):
-        _run_step_migration(self._p1_handler, step)
-        assert Path(f"{step_dir}/dir1/foo").exists()
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-        assert Path(f"overlay/{step_dir}_overlay").exists()
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_migrate_overlay_whiteout_translation(
-        self, mocker, new_dir, step, step_dir
-    ):
-        wh = Path("parts/p2/layer/dir1/foo")
-        wh.touch()
-        mocker.patch(
-            "craft_parts.overlays.is_whiteout_file", new=lambda x: x == new_dir / wh
-        )
-
-        _run_step_migration(self._p1_handler, step)
-        assert Path(f"{step_dir}/dir1/foo").exists() is False
-        assert Path(f"{step_dir}/dir1/.wh.foo").exists()
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_migrate_overlay_opaque_dir_translation(
-        self, mocker, new_dir, step, step_dir
-    ):
-        opaque = Path("parts/p2/layer/dir1")
-        mocker.patch(
-            "craft_parts.overlays.is_opaque_dir", new=lambda x: x == new_dir / opaque
-        )
-
-        _run_step_migration(self._p1_handler, step)
-        assert Path(f"{step_dir}/dir1/.wh..wh..opq").exists()
-        assert Path(f"{step_dir}/dir1/foo").exists() is False
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_clean_migrated_overlay(self, mocker, new_dir, step, step_dir):
-        wh = Path("parts/p2/layer/dir1/foo")
-        wh.touch()
-        mocker.patch(
-            "craft_parts.overlays.is_whiteout_file", new=lambda x: x == new_dir / wh
-        )
-
-        opaque = Path("parts/p2/layer/dir2")
-        opaque.mkdir()
-        mocker.patch(
-            "craft_parts.overlays.is_opaque_dir", new=lambda x: x == new_dir / opaque
-        )
-
-        _run_step_migration(self._p1_handler, step)
-        assert Path(f"{step_dir}/dir1/foo").exists() is False
-        assert Path(f"{step_dir}/dir1/.wh.foo").exists()
-        assert Path(f"{step_dir}/dir2/.wh..wh..opq").exists()
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-        assert Path(f"overlay/{step_dir}_overlay").exists()
-
-        self._p1_handler.clean_step(step)
-        assert Path(f"{step_dir}/dir1/.wh.foo").exists() is False
-        assert Path(f"{step_dir}/dir2").exists() is False
-        assert Path(f"{step_dir}/bar").exists() is False
-        assert Path(f"{step_dir}/dir1/baz").exists() is False
-        assert Path(f"overlay/{step_dir}_overlay").exists() is False
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_clean_stage_overlay_multiple_parts(self, step, step_dir):
-        _run_step_migration(self._p1_handler, step)
-        assert Path(f"{step_dir}/dir1/foo").exists()
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-        assert Path(f"overlay/{step_dir}_overlay").exists()
-
-        _run_step_migration(self._p2_handler, step)
-
-        self._p1_handler.clean_step(step)
-        assert Path(f"{step_dir}/dir1/foo").exists()
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/dir1/baz").exists()
-        assert Path(f"overlay/{step_dir}_overlay").exists()
-
-        self._p2_handler.clean_step(step)
-        assert Path(f"{step_dir}/dir1/foo").exists() is False
-        assert Path(f"{step_dir}/bar").exists() is False
-        assert Path(f"{step_dir}/dir1/baz").exists() is False
-        assert Path(f"overlay/{step_dir}_overlay").exists() is False
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_clean_overlay_shared_file(self, mocker, step, step_dir):
-        Path("parts/p1/layer/file1").write_text("content")
-        Path("parts/p3/install/file1").write_text("content")
-
-        _run_step_migration(self._p2_handler, step)
-        _run_step_migration(self._p3_handler, step)
-        assert Path(f"{step_dir}/file1").exists()
-        assert Path(f"{step_dir}/bar").exists()
-
-        # clean overlay data
-        self._p1_handler.clean_step(step)
-        assert Path(f"{step_dir}/bar").exists() is False
-        assert Path(f"{step_dir}/file1").exists()  # file1 remains (also belongs to p3)
-
-        # clean part data
-        self._p3_handler.clean_step(step)
-        assert Path(f"{step_dir}/file1").exists() is False
-
-    @pytest.mark.parametrize(
-        "step,step_dir", [(Step.STAGE, "stage"), (Step.PRIME, "prime")]
-    )
-    def test_clean_part_shared_file(self, mocker, step, step_dir):
-        Path("parts/p1/layer/file1").write_text("content")
-        Path("parts/p3/install/file1").write_text("content")
-
-        _run_step_migration(self._p2_handler, step)
-        _run_step_migration(self._p3_handler, step)
-        assert Path(f"{step_dir}/file1").exists()
-        assert Path(f"{step_dir}/bar").exists()
-
-        # clean part data
-        self._p3_handler.clean_step(step)
-        assert Path(f"{step_dir}/bar").exists()
-        assert Path(f"{step_dir}/file1").exists()  # file1 remains (also belongs to p1)
-
-        # clean overlay data
-        self._p1_handler.clean_step(step)
-        assert Path(f"{step_dir}/file1").exists() is False
-
-    def test_migrate_overlay_filter_whiteout(self, mocker, new_dir):
-        cache_dir = new_dir / "cache"
-        base_dir = new_dir / "base"
-        cache_dir.mkdir()
-        base_dir.mkdir()
-
-        p1 = Part("p1", {"plugin": "nil", "overlay-script": "ls"})
-        info = ProjectInfo(application_name="test", cache_dir=cache_dir)
-        ovmgr = OverlayManager(
-            project_info=info, part_list=[p1], base_layer_dir=base_dir
-        )
-        p1_handler = PartHandler(
-            p1,
-            part_info=PartInfo(info, p1),
-            part_list=[p1],
-            overlay_manager=ovmgr,
-        )
-        p1_handler._make_dirs()
-
-        wh1 = p1.part_layer_dir / "file1"
-        wh2 = p1.part_layer_dir / "file2"
-        wh1.touch()
-        wh2.touch()
-        mocker.patch(
-            "craft_parts.overlays.is_whiteout_file", new=lambda x: x in (wh1, wh2)
-        )
-
-        Path(base_dir, "file2").touch()  # file2 has a backing file
-
-        _run_step_migration(p1_handler, Step.PRIME)
-        assert Path("prime/.wh.file1").exists() is False
-        assert Path("prime/.wh.file2").exists()
 
 
 def _run_step_migration(handler: PartHandler, step: Step) -> None:
@@ -1166,9 +811,8 @@ class TestPackages:
         assert handler.build_snaps == ["word-salad"]
 
 
-@pytest.mark.usefixtures("new_dir")
 class TestFileFilter:
-    """Overlay filter test cases."""
+    """File filter test cases."""
 
     @pytest.fixture(autouse=True)
     def setup_method_fixture(self, new_dir):
@@ -1183,8 +827,8 @@ class TestFileFilter:
 
     def test_apply_file_filter_empty(self, new_dir):
         destdir = Path("destdir")
-        overlay_fileset = filesets.Fileset([])
-        files, dirs = filesets.migratable_filesets(overlay_fileset, str(destdir))
+        fileset = filesets.Fileset([])
+        files, dirs = filesets.migratable_filesets(fileset, str(destdir))
         part_handler._apply_file_filter(
             filter_files=files, filter_dirs=dirs, destdir=destdir
         )
@@ -1198,8 +842,8 @@ class TestFileFilter:
 
     def test_apply_file_filter_remove_file(self, new_dir):
         destdir = Path("destdir")
-        overlay_fileset = filesets.Fileset(["-file1", "-dir1/file3"])
-        files, dirs = filesets.migratable_filesets(overlay_fileset, str(destdir))
+        fileset = filesets.Fileset(["-file1", "-dir1/file3"])
+        files, dirs = filesets.migratable_filesets(fileset, str(destdir))
         part_handler._apply_file_filter(
             filter_files=files, filter_dirs=dirs, destdir=destdir
         )
@@ -1213,8 +857,8 @@ class TestFileFilter:
 
     def test_apply_file_filter_remove_dir(self, new_dir):
         destdir = Path("destdir")
-        overlay_fileset = filesets.Fileset(["-dir1", "-dir1/dir2"])
-        files, dirs = filesets.migratable_filesets(overlay_fileset, str(destdir))
+        fileset = filesets.Fileset(["-dir1", "-dir1/dir2"])
+        files, dirs = filesets.migratable_filesets(fileset, str(destdir))
         part_handler._apply_file_filter(
             filter_files=files, filter_dirs=dirs, destdir=destdir
         )
@@ -1227,8 +871,8 @@ class TestFileFilter:
 
     def test_apply_file_filter_remove_symlink(self, new_dir):
         destdir = Path("destdir")
-        overlay_fileset = filesets.Fileset(["-file4", "-dir3"])
-        files, dirs = filesets.migratable_filesets(overlay_fileset, str(destdir))
+        fileset = filesets.Fileset(["-file4", "-dir3"])
+        files, dirs = filesets.migratable_filesets(fileset, str(destdir))
         part_handler._apply_file_filter(
             filter_files=files, filter_dirs=dirs, destdir=destdir
         )
@@ -1242,8 +886,8 @@ class TestFileFilter:
 
     def test_apply_file_filter_keep_file(self, new_dir):
         destdir = Path("destdir")
-        overlay_fileset = filesets.Fileset(["dir1/file3"])
-        files, dirs = filesets.migratable_filesets(overlay_fileset, str(destdir))
+        fileset = filesets.Fileset(["dir1/file3"])
+        files, dirs = filesets.migratable_filesets(fileset, str(destdir))
         part_handler._apply_file_filter(
             filter_files=files, filter_dirs=dirs, destdir=destdir
         )
@@ -1287,21 +931,3 @@ class TestHelpers:
     def test_remove_non_existent(self):
         # this should not raise and exception
         part_handler._remove(Path("not_here"))
-
-    @pytest.mark.parametrize("step", list(Step))
-    def test_parts_with_overlay_in_step(self, step):
-        p1 = Part("p1", {"plugin": "nil"})
-        p2 = Part("p2", {"plugin": "nil", "overlay-script": "ls"})
-        p3 = Part("p3", {"plugin": "nil", "overlay-packages": ["pkg1"]})
-        p4 = Part("p4", {"plugin": "nil", "overlay": ["/etc"]})
-
-        res = part_handler._parts_with_overlay_in_step(step, part_list=[p1, p2, p3, p4])
-        assert res == []
-
-        for part in [p1, p2, p3, p4]:
-            state_path = states.get_step_state_path(part, step)
-            state_path.parent.mkdir(parents=True)
-            state_path.touch()
-
-        res = part_handler._parts_with_overlay_in_step(step, part_list=[p1, p2, p3, p4])
-        assert res == [p2, p3, p4]
