+++ tests/unit/features/overlay/test_executor_environment.py
@@ -104,6 +104,7 @@ def test_generate_step_environment_build(new_dir):
         export CRAFT_TARGET_ARCH="arm64"
         export CRAFT_PARALLEL_BUILD_COUNT="1"
         export CRAFT_PROJECT_DIR="{new_dir}"
+        export CRAFT_OVERLAY="{new_dir}/overlay/overlay"
         export CRAFT_STAGE="{new_dir}/stage"
         export CRAFT_PRIME="{new_dir}/prime"
         export CRAFT_PROJECT_NAME="test-project"
@@ -152,6 +153,7 @@ def test_generate_step_environment_no_project_name(new_dir):
         export CRAFT_TARGET_ARCH="arm64"
         export CRAFT_PARALLEL_BUILD_COUNT="1"
         export CRAFT_PROJECT_DIR="{new_dir}"
+        export CRAFT_OVERLAY="{new_dir}/overlay/overlay"
         export CRAFT_STAGE="{new_dir}/stage"
         export CRAFT_PRIME="{new_dir}/prime"
         export CRAFT_PART_NAME="p1"
@@ -202,6 +204,7 @@ def test_generate_step_environment_no_build(new_dir, step):
         export CRAFT_TARGET_ARCH="arm64"
         export CRAFT_PARALLEL_BUILD_COUNT="1"
         export CRAFT_PROJECT_DIR="{new_dir}"
+        export CRAFT_OVERLAY="{new_dir}/overlay/overlay"
         export CRAFT_STAGE="{new_dir}/stage"
         export CRAFT_PRIME="{new_dir}/prime"
         export CRAFT_PROJECT_NAME="test-project"
@@ -251,6 +254,7 @@ def test_generate_step_environment_no_user_env(new_dir):
         export CRAFT_TARGET_ARCH="arm64"
         export CRAFT_PARALLEL_BUILD_COUNT="1"
         export CRAFT_PROJECT_DIR="{new_dir}"
+        export CRAFT_OVERLAY="{new_dir}/overlay/overlay"
         export CRAFT_STAGE="{new_dir}/stage"
         export CRAFT_PRIME="{new_dir}/prime"
         export CRAFT_PROJECT_NAME="test-project"
